Move warnings and alerts to a separate directory.
