+++ src/core/warnings/Warning.js
